requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
faiss-cpu==1.7.4
openai==0.28.0
python-dotenv==1.0.0
PyPDF2==3.0.1
numpy
pydantic==2.0
python-multipart==0.0.6